Add user database file to data folder and remove user database from express_server.js
