+++ data/user-database.js
@@ -1,0 +1,20 @@
+// User Database
+const users = {
+  "userRandomID": {
+    id: "userRandomID",
+    email: "user@example.com",
+    password: "purple-monkey-dinosaur"
+  },
+  "user2RandomID": {
+    id: "user2RandomID",
+    email: "user2@example.com",
+    password: "dishwasher-funk"
+  },
+  '80e100': {
+    id: '80e100',
+    email: 'sribas@gmail.com',
+    password: '$2b$10$2dA3e5.e3H51UTayA30qruiQrNw73lBfq474YfiTltNLPinfzWWVq'
+  }
+};
+
+module.exports = { users };
